create `views.py` module
